clean up rendering
fixes #5
- refactor rendering
- minify html
