+++ package.json
@@ -23,11 +23,13 @@
     "dependencies": {
         "@sendgrid/mail": "^6.3.1",
         "@types/csvtojson": "^1.1.5",
+        "@types/html-minifier": "^3.5.2",
         "@types/node": "^10.12.18",
         "@types/puppeteer": "^1.11.1",
         "@types/shelljs": "^0.8.1",
         "csvtojson": "^2.0.8",
         "get-cursor-position": "^1.0.5",
+        "html-minifier": "^3.5.21",
         "puppeteer": "^1.11.0",
         "shelljs": "^0.8.3"
     }
