+++ src/spendsummary.d.ts
@@ -1,0 +1,5 @@
+interface SpendSummary {
+    name: string;
+    total: string;
+    count: number;
+}
